Revert "Improving code"
