--- .vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "go.inferGopath": false
-}
