--- data/data.go
@@ -1,52 +1,0 @@
-package data
-
-func parse_json() []string {
-	listname := []string{
-		"Amplitude API Key",
-		"Asana Access token",
-		"Bing Maps API Key",
-		"Bit.ly Access token",
-		"Block.io API Key",
-		"Branch.IO Key and Secret",
-		"BrowserStack Access Key",
-		"Buildkite Access token",
-		"ButterCMS API Key",
-		"Calendarific API Key",
-		"Calendly API Key",
-		"CircleCI Access Token",
-		"Coinmarketcap API Key",
-		"Covalent API Key",
-		"Cryptocompare API Key",
-		"DataDog API key",
-		"EAN-Search Access Token",
-		"Facebook Access Token",
-		"Fungenerator API Key",
-		"Github Token",
-		"Heroku API Key",
-		"Holiday API Key",
-		"Hubspot API Key",
-		"Ipstack API Key",
-		"Iterable API Key",
-		"Jumpcloud API key",
-		"Localytics API and Secret key",
-		"Lokalise API Key",
-		"Loqate API key",
-		"MailGun API Key",
-		"Mapbox API Key",
-		"Pagerduty API token",
-		"Pendo Integration Key",
-		"PivotalTracker API Token",
-		"Razorpay API key and secret key",
-		"SauceLabs Username and access Key",
-		"SendGrid API Token",
-		"Spotify Access Token",
-		"Stripe Live Token",
-		"Travis CI API token",
-		"Twilio Account_sid and Auth token",
-		"Visual Studio App Center API Token",
-		"WPEngine API Key",
-		"WakaTime API Key",
-		"Youtube API Key",
-		"Zendesk Access token"}
-	return listname
-}
